just renamed exerciseTracker Module to userExerciseModule
