+++ the-pack-backend/src/userExerciseTracker/userExerciseTracker.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
-import { ExerciseTrackerService } from './exerciseTracker.service';
+import { ExerciseTrackerService } from './userExerciseTracker.service';
 import { Prisma } from '@prisma/client';
 
 @Controller('exerciseTracker')
