+++ the-pack-backend/src/userExerciseTracker/userExerciseTracker.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { ExerciseTrackerController } from './exerciseTracker.controller';
-import { ExerciseTrackerService } from './exerciseTracker.service';
+import { ExerciseTrackerController } from './userExerciseTracker.controller';
+import { ExerciseTrackerService } from './userExerciseTracker.service';
 import { PrismaModule } from 'src/prisma/prisma.module';
 import { PrismaService } from 'src/prisma/prisma.service';
